requirements.txt
Django>=5.0
djangorestframework>=3.15
requests
beautifulsoup4
pandas
schedule
certifi
python-certifi-win32
